pyproject.toml
[project]
name = "syscleaner"
version = "0.2.0"
description = "Универсальный инструмент для очистки и аудита системы (macOS и Linux)"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Aleksandr Mordvinov"}
]
keywords = [
    "system-cleaner",
    "macos",
    "linux",
    "cleanup",
    "audit",
    "security",
    "ml-cache",
    "huggingface",
    "pytorch",
    "dependencies",
    "dependency-analysis",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "tomli>=2.0.0; python_version < '3.11'",
]

[project.urls]
Homepage = "https://github.com/FUYOH666/Cleaner-OS"
Repository = "https://github.com/FUYOH666/Cleaner-OS"
Issues = "https://github.com/FUYOH666/Cleaner-OS/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
    "pyright>=1.1.0",
]

[project.scripts]
syscleaner = "syscleaner.main:app"
system-cleaner = "syscleaner.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.pyright]
pythonVersion = "3.12"
pythonPlatform = ["Darwin", "Linux"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

